requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
authlib==1.3.2
httpx==0.27.2
python-jose==3.3.0
pydantic==2.9.2
jinja2==3.1.4
itsdangerous==2.2.0
python-dotenv==1.0.1
